+++ src/lib.rs
@@ -3,10 +3,10 @@ mod event;
 mod filter;
 mod node;
 mod query;
-mod tree;
+mod state;
 
 pub use event::*;
 pub use filter::*;
 pub use node::*;
 pub use query::*;
-pub use tree::*;
+pub use state::*;
